A few edits to pregnancies
Edit pregnancy metric tracking to add when getting pregnant, then check
for knowledge when showing it. Makes more sense to me this way, and
fixes a usage of it when milking in the farm. Will also count
pregnancies that are aborted before being known.

Fix a condition using .append('currentpregnancy') instead of .has

Change order of impregnation and ovulation processing, so the ovulation
state the player sees is accurate.