Cache images for performance
It takes a moderate amount of time to load an image from a file, which
was being done a lot with the same images, specifically at end of day
when updating body image for a lot of slaves. Instead of loading these
images every time, make a cache for them.