Image loading performance tweaks
Don't actually load the image in canloadimage(), since that can be slow
when called a lot. Will change behavior in a couple places if the file
exists but can't be loaded, showing a blank panel instead it being
hidden.

Edit Mansion slave list updating to not reload images that didn't
change.